Version 1.0

+++ back-end/cinema-api/cinema-api/Data/SqlSectionRepo.cs
@@ -1,4 +1,5 @@
 ﻿using cinema_api.Model;
+using Microsoft.EntityFrameworkCore;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -26,7 +27,12 @@ namespace cinema_api.Data
 
         public void DeleteSection(Section section)
         {
-            if(DateTime.Now.AddDays(10) <= section.Date || DateTime.Now > section.Date)
+            if (section == null)
+            {
+                throw new ArgumentNullException(nameof(section));
+            }
+
+            if (DateTime.Now.AddDays(10) <= section.Date || DateTime.Now > section.Date)
             {
                 _context.Sections.Remove(section);
             }
@@ -39,12 +45,13 @@ namespace cinema_api.Data
 
         public IEnumerable<Section> GetAllSections()
         {
-            return _context.Sections.ToList().OrderBy(p => p.Date);
+            return _context.Sections.Include(p => p.Film).Include(p=> p.Room).ToList().OrderBy(p => p.Date);
         }
 
         public Section GetSectionById(int Id)
         {
-            return _context.Sections.FirstOrDefault(p => p.Id == Id);
+            return _context.Sections.Include(p => p.Film).Include(p => p.Room)
+                .FirstOrDefault(p => p.Id == Id);
         }
 
         public bool SaveChanges()
@@ -55,6 +62,29 @@ namespace cinema_api.Data
         public void UpdateSection(Section section)
         {
             
+        }
+
+        public bool ValidateSection(Section sectionModel)
+        {
+            var sections = _context.Sections.ToList();
+
+            foreach (var section in sections)
+            {
+                if(section.Room == sectionModel.Room)
+                {
+                    var a = section.Film.Duration.Split(':');
+                    var b = sectionModel.Film.Duration.Split(':');
+
+                    var aux1 = (Convert.ToInt32(a[0]) * 60) + Convert.ToInt32(a[1]);
+                    var aux2 = (Convert.ToInt32(b[0]) * 60) + Convert.ToInt32(b[1]);
+                    if (section.Date.AddMinutes(aux1) >= sectionModel.Date || sectionModel.Date.AddMinutes(aux2) >= section.Date)
+                    {
+                        return false;
+                    }
+                }
+            }
+
+            return true;
         }
     }
 }
